backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
pdfminer.six
docx2txt
